refactor: move ssr to bindings
