+++ bindings/gpt-vis-ssr/.fatherrc.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'father';
+
+export default defineConfig({
+  esm: {
+    output: 'dist/esm',
+  },
+  cjs: {
+    output: 'dist/cjs',
+  },
+});
