+++ bindings/gpt-vis-ssr/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "@antv/gpt-vis-ssr",
+  "version": "0.1.0",
+  "description": "SSR(Server Side Render) for GPT-Vis.",
+  "keywords": [
+    "antv",
+    "GPTs",
+    "ssr",
+    "vis",
+    "GPT-Vis"
+  ],
+  "homepage": "https://gpt-vis.antv.vision",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/antvis/GPT-Vis"
+  },
+  "license": "MIT",
+  "author": "antvis",
+  "sideEffects": false,
+  "type": "module",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "typings": "dist/esm/index.d.ts",
+  "files": [
+    "src"
+  ],
+  "scripts": {
+    "build": "father build",
+    "check-deps": "father doctor",
+    "test:unit": "jest --config ./jest.config.ts"
+  },
+  "dependencies": {
+    "@antv/g2-ssr": "^0.0.8",
+    "@antv/g6-ssr": "^0.0.16",
+    "@antv/gpt-vis": "^0.4.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.14",
+    "father": "^4.5.2",
+    "jest": "^29.7.0",
+    "npm-run-all": "^4.1.5",
+    "ts-jest": "^29.3.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3",
+    "canvas": "^3.1.0"
+  },
+  "peerDependencies": {
+    "canvas": "^3.1.0"
+  },
+  "engines": {
+    "node": ">=18",
+    "pnpm": ">=8"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org"
+  }
+}
