+++ bindings/gpt-vis-ssr/tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "compilerOptions": {
+    "target": "esnext",
+    "module": "esnext",
+    "moduleResolution": "node",
+    "jsx": "react",
+    "esModuleInterop": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "declaration": true,
+    "resolveJsonModule": true,
+    "baseUrl": "./",
+    "paths": {
+      "@antv/gpt-vis": ["./src/index.ts"]
+    }
+  },
+  "include": ["./*.?*.ts", "src", "__tests__", "docs"],
+  "exclude": ["node_modules", "dist"]
+}
